# openapi.yaml
openapi: 3.0.3
info:
  title: ias_stt — Trap Management API
  version: 1.0.0
  description: |
    Trap management service with deployment tracking, location
    history, and immutable deployment actions with required pictures. Telemetry is ingested from the
    ChirpStack HTTP integration via `POST /api/telemetry/ingest`
    (an optional MQTT listener remains available). Application users obtain
    JWT access and refresh tokens from `/auth/login` and send access tokens in
    the `Authorization` header. A legacy API key remains available for service
    integrations.
servers:
  - url: http://localhost:8080
    description: Local development
security:
  - bearerAuth: []
...
paths:
...
  /api/ldap/users:
    get:
      tags: [Users]
      summary: Search directory identities for administrator provisioning
      operationId: searchDirectoryUsers
      description: Administrator-only. Returns profile fields and a stable opaque directory subject; never returns a DN or credentials.
      parameters:
        - name: search
          in: query
          required: true
          schema: { type: string, minLength: 2 }
        - name: limit
          in: query
          schema: { type: integer, minimum: 1, maximum: 50, default: 25 }
        - name: offset
          in: query
          schema: { type: integer, minimum: 0, default: 0 }
      responses:
        200:
          description: Directory search results
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/DirectoryUser' }
        401:
          $ref: '#/components/responses/401'
        403:
          $ref: '#/components/responses/403'
        422:
          $ref: '#/components/responses/422'
        503:
          $ref: '#/components/responses/503'
...
  /api/users:
    get:
      tags: [Users]
      summary: List provisioned application users
      operationId: listUsers
      parameters:
        - $ref: '#/components/parameters/limit'
        - $ref: '#/components/parameters/offset'
        - name: search
          in: query
          schema: { type: string }
          description: Search username, display name, or email
      responses:
        200:
          description: Users array
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/User' }
        401:
          $ref: '#/components/responses/401'
        403:
          $ref: '#/components/responses/403'
        422:
          $ref: '#/components/responses/422'

    post:
      tags: [Users]
      summary: Create a LOCAL user or provision an LDAP user
      operationId: createUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserCreate'
      responses:
        201:
          description: User created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        401:
          $ref: '#/components/responses/401'
        403:
          $ref: '#/components/responses/403'
        404:
          $ref: '#/components/responses/404'
        409:
          $ref: '#/components/responses/409'
        422:
          $ref: '#/components/responses/422'
        503:
          $ref: '#/components/responses/503'
  /api/users/{user_id}:
    parameters:
      - name: user_id
        in: path
        required: true
        schema: { type: integer }
    get:
      tags: [Users]
      summary: Get an application user
      operationId: getUser
      responses:
        200:
          description: User found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        401:
          $ref: '#/components/responses/401'
        403:
          $ref: '#/components/responses/403'
        404:
          $ref: '#/components/responses/404'

    put:
      tags: [Users]
      summary: Update application-managed user fields
      operationId: updateUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserUpdate'
      responses:
        200:
          description: User updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        401:
          $ref: '#/components/responses/401'
        403:
          $ref: '#/components/responses/403'
        404:
          $ref: '#/components/responses/404'
        409:
          $ref: '#/components/responses/409'
        422:
          $ref: '#/components/responses/422'
  /api/users/{user_id}/sync-profile:
    parameters:
      - name: user_id
        in: path
        required: true
        schema: { type: integer }
    post:
      tags: [Users]
      summary: Synchronize an LDAP user's profile
      operationId: syncLdapUserProfile
      description: Administrator-only. Updates directory-owned username, display name, and email; local role and enabled status are preserved.
      responses:
        200:
          description: Synchronized application user
          content:
            application/json:
              schema: { $ref: '#/components/schemas/User' }
        401:
          $ref: '#/components/responses/401'
        403:
          $ref: '#/components/responses/403'
        404:
          $ref: '#/components/responses/404'
        409:
          $ref: '#/components/responses/409'
        503:
          $ref: '#/components/responses/503'

    delete:
      tags: [Users]
      summary: Delete a local user
      operationId: deleteUser
      responses:
        200:
          description: User deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  message: { type: string }
        401:
          $ref: '#/components/responses/401'
        403:
          $ref: '#/components/responses/403'
        404:
          $ref: '#/components/responses/404'
        409:
          $ref: '#/components/responses/409'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: |
        Application users pass the JWT access token as
        `Authorization: Bearer <access-token>`. Legacy service integrations
        may pass the configured API key in the same header.

  # -- Reusable parameters ---------------------------------------------------
  parameters:
    limit:
      name: limit
      in: query
      schema: { type: integer, default: 100, minimum: 0 }
      description: Maximum items per page
...
    offset:
      name: offset
      in: query
      schema: { type: integer, default: 0, minimum: 0 }
      description: Number of items to skip

  # -- Schemas ---------------------------------------------------------------
  schemas:
    Error:
      type: object
      properties:
        error:
          oneOf:
            - type: string
            - type: object
              properties:
                code: { type: string }
                message: { type: string }
      example:
        error:
          code: AUTHENTICATION_REQUIRED
          message: Authentication is required.
...
    DirectoryUser:
      type: object
      properties:
        directory_key: { type: string }
        directory_subject: { type: string, description: Stable directory identity used for provisioning. }
        username: { type: string }
        display_name: { type: string, nullable: true }
        email: { type: string, nullable: true }
        is_enabled: { type: boolean, nullable: true }
...
    User:
      type: object
      properties:
        id: { type: integer, readOnly: true }
        username: { type: string, maxLength: 150, readOnly: true }
        auth_provider: { type: string, enum: [LOCAL, LDAP], readOnly: true }
        directory_key: { type: string, maxLength: 100, nullable: true, readOnly: true }
        directory_subject: { type: string, maxLength: 255, nullable: true, readOnly: true }
        ldap_synced_at: { type: string, format: date-time, nullable: true, readOnly: true }
        display_name: { type: string, maxLength: 255, nullable: true }
        email: { type: string, maxLength: 255, nullable: true }
        role:
          type: string
          enum: [administrator, field_operator, read_only]
        is_active: { type: boolean }
        created_at: { type: string, format: date-time, readOnly: true }
        updated_at: { type: string, format: date-time, readOnly: true }
    UserCreate:
      type: object
      description: Supply the fields appropriate to the selected provider. LDAP profile data is resolved by the backend from directory_subject.
      required: [auth_provider, role]
      properties:
        auth_provider: { type: string, enum: [LOCAL, LDAP] }
        username: { type: string, maxLength: 150 }
        password: { type: string, format: password, minLength: 8, writeOnly: true }
        directory_subject: { type: string, maxLength: 255, writeOnly: true }
        display_name: { type: string, maxLength: 255, nullable: true }
        email: { type: string, maxLength: 255, nullable: true }
        role:
          type: string
          enum: [administrator, field_operator, read_only]
        is_active: { type: boolean, default: true }
    UserUpdate:
      type: object
      properties:
        password: { type: string, format: password, minLength: 8, writeOnly: true }
        display_name: { type: string, maxLength: 255, nullable: true }
        email: { type: string, maxLength: 255, nullable: true }
        role:
          type: string
          enum: [administrator, field_operator, read_only]
        is_active: { type: boolean }
...
  responses:
...
    401:
      description: Missing or invalid JWT or service credential
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error:
              code: AUTHENTICATION_REQUIRED
              message: Authentication is required.
    403:
      description: Authenticated user lacks the required permission
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error:
              code: FORBIDDEN
              message: You do not have permission to perform this action.
    422:
      description: Validation error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error:
              code: VALIDATION_ERROR
              message: Field 'password' must be at least 8 characters
    404:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: Trap not found
    409:
      description: Conflict — duplicate trap_id
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: trap_id 'TRAP-001' already exists
...
    503:
      description: The configured LDAP authentication provider is unavailable or misconfigured
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error:
              code: AUTH_PROVIDER_UNAVAILABLE
              message: The authentication service is temporarily unavailable.